backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Async & Task Queue
celery==5.3.4
redis==5.0.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# AI/ML & NLP
transformers==4.35.2
torch==2.1.1
bertopic==0.16.0
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
sentence-transformers==2.2.2

# External AI APIs
twelvelabs-python==0.2.0
openai==1.3.7
elevenlabs==0.2.26

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AWS Integration
boto3==1.29.7
botocore==1.32.7

# Export & Reporting
reportlab==4.0.7
python-pptx==0.6.23
jinja2==3.1.2

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
tenacity==8.2.3
python-dateutil==2.8.2

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
